Refactor provider references and streamFactory
Moved `Providers` and `ApiKeys` from `plugin.ts` to a new file `providers.ts` for better organization. This change impacts how streams are fetched and accessed.
